+++ dvha/tools/dicom_dose_sum.py
@@ -12,7 +12,7 @@ Class for summing dose grids
 import numpy as np
 from os.path import isfile
 import pydicom
-from scipy.interpolate import RegularGridInterpolator
+from scipy.ndimage import map_coordinates
 from copy import deepcopy
 
 
@@ -27,10 +27,12 @@ class DoseGrid:
         grid_sum.save_dcm(some_file_path)
 
     """
-    def __init__(self, rt_dose, try_full_interp=True, interp_block_size=50000):
+    def __init__(self, rt_dose, order=1, try_full_interp=True, interp_block_size=50000):
         """
         :param rt_dose: an RT Dose DICOM dataset or file_path
         :type rt_dose: pydicom.FileDataset
+        :param order: The order of the spline interpolation. The order has to be in the range 0-5.
+                      0: the nearest grid point, 1: trilinear, 2 to 5: spline
         :param try_full_interp: If true, will attempt to interpolate the entire grid at once before calculating one
                                 block at a time (block size defined in self.interp_by_block)
         :type try_full_interp: bool
@@ -39,6 +41,7 @@ class DoseGrid:
         """
 
         self.ds = self.__validate_input(rt_dose)
+        self.order = order
         self.try_full_interp = try_full_interp
         self.interp_block_size = interp_block_size
 
@@ -135,6 +138,17 @@ class DoseGrid:
         """Save the pydicom.FileDataset to file"""
         self.ds.save_as(file_path)
 
+    def get_ijk_points(self, other_axes):
+        """
+        Convert axes from another DoseGrid into ijk of this DoseGrid
+        :param other_axes: the x, y, and z axis arrays
+        :type other_axes: list
+        :return: np.vstack of other_axes in this ijk space
+        """
+        ijk_axes = [(np.array(axis) - self.offset[a]) / self.scale[a] for a, axis in enumerate(other_axes)]
+        j, i, k = np.meshgrid(ijk_axes[1], ijk_axes[0], ijk_axes[2])
+        return np.vstack((i.ravel(), j.ravel(), k.ravel()))
+
     ####################################################
     # Dose Summation
     ####################################################
@@ -160,35 +174,35 @@ class DoseGrid:
         :param other: another DoseGrid
         :type other: DoseGrid
         """
-
-        # TODO: Try scipy.ndimage.map_coordinates again, may be faster?
-        interpolator = RegularGridInterpolator(points=other.axes, values=other.dose_grid,
-                                               bounds_error=False, fill_value=0)
-
         other_grid = None
         if self.try_full_interp:
             try:
-                other_grid = self.interp_entire_grid(interpolator)
-            except MemoryError as e:
+                other_grid = self.interp_entire_grid(other)
+            except MemoryError:
                 pass
         if other_grid is None:
-            other_grid = self.interp_by_block(interpolator)
+            other_grid = self.interp_by_block(other)
 
         self.dose_grid += other_grid
         self.set_pixel_data()
 
-    def interp_entire_grid(self, interpolator):
-        """Interpolate the other dose grid to this dose grid's axes"""
-        return interpolator(self.points).reshape(self.shape)
+    def interp_entire_grid(self, other):
+        """
+        Interpolate the other dose grid to this dose grid's axes
+        :param other: another DoseGrid
+        :type other: DoseGrid
+        """
+        points = other.get_ijk_points(self.axes)
+        return map_coordinates(input=other.dose_grid, coordinates=points, order=self.order).reshape(self.shape)
 
-    def interp_by_block(self, interpolator):
+    def interp_by_block(self, other):
         """
         Interpolate the other dose grid to this dose grid's axes, calculating one block at a time
         The block is defined at the init of this class, default is 50,000 points at a time
-        :param interpolator: object to perform interpolation
-        :type interpolator: RegularGridInterpolator
+        :param other: another DoseGrid
+        :type other: DoseGrid
         """
-        points = self.points
+        points = other.get_ijk_points(self.axes)
         point_count = np.product(self.shape)
         other_grid = np.zeros(point_count)
 
@@ -197,6 +211,7 @@ class DoseGrid:
         for i in range(block_count):
             start = i * self.interp_block_size
             end = (i+1) * self.interp_block_size if i + 1 < block_count else -1
-            other_grid[start:end] = interpolator(points[start:end])
+            other_grid[start:end] = map_coordinates(input=other.dose_grid, coordinates=points[start:end],
+                                                    order=self.order)
 
         return other_grid.reshape(self.shape)